[ Add ] : TimeInterval extensions
